+++ apps/block_scout_web/lib/block_scout_web/controllers/address_controller.ex
@@ -60,7 +60,7 @@ defmodule BlockScoutWeb.AddressController do
   def index(conn, _params) do
     render(conn, "index.html",
       current_path: current_path(conn),
-      address_count: Chain.count_addresses_from_cache()
+      address_count: Chain.address_estimated_count()
     )
   end