+++ apps/block_scout_web/test/block_scout_web/features/viewing_addresses_test.exs
@@ -3,7 +3,6 @@ defmodule BlockScoutWeb.ViewingAddressesTest do
     # Because ETS tables is shared for `Explorer.Counters.*`
     async: false
 
-  alias Explorer.Counters.AddressesCounter
   alias BlockScoutWeb.{AddressPage, AddressView, Notifier}
 
   setup do
@@ -57,9 +56,6 @@ defmodule BlockScoutWeb.ViewingAddressesTest do
     test "lists top addresses", %{session: session, addresses: addresses} do
       [first_address | _] = addresses
       [last_address | _] = Enum.reverse(addresses)
-
-      start_supervised!(AddressesCounter)
-      AddressesCounter.consolidate()
 
       session
       |> AddressPage.visit_page()