+++ apps/block_scout_web/test/block_scout_web/features/viewing_app_test.exs
@@ -5,14 +5,10 @@ defmodule BlockScoutWeb.ViewingAppTest do
 
   alias BlockScoutWeb.AppPage
   alias BlockScoutWeb.Counters.BlocksIndexedCounter
-  alias Explorer.Counters.AddressesCounter
   alias Explorer.{Repo}
   alias Explorer.Chain.{Transaction}
 
   setup do
-    start_supervised!(AddressesCounter)
-    AddressesCounter.consolidate()
-
     :ok
   end